# Speed up parsing with many namespaces