@w666 2025-08-03
Hi @nora-heydecker-axa, thanks for the PR.
Interesting results. Changes look good to me. Please fix linter issue and make sure tests pass.
You can do it locally
```
npm i
npm build
npm test
```