@nora-heydecker-axa 2025-08-04
I have run the tests and linter locally and fixed the missing semicolon.