@w666 2025-08-05
Interesting, I've noticed that tests 2 times slower now, from 2-3 seconds up to 4-5 seconds on my machine.
I will double check this before release.