@w666 2025-08-13
I released 1.3.0 earlier today. Thanks for the contribution.